Split out an inner class that was duplicated; put the tests in groups; get rid of AAPB reference.
