+++ lib/sony-ci-api/sony_ci_basic.rb
@@ -1,6 +1,7 @@
 require 'yaml'
 require 'curb'
 require 'json'
+require_relative 'sony_ci_client'
 
 class SonyCiBasic
   attr_reader :access_token
@@ -50,23 +51,7 @@ class SonyCiBasic
     Downloader.new(self).download(asset_id)
   end
 
-  class CiClient #:nodoc:
-    # This class hierarchy might be excessive, but it gives us:
-    # - a single place for the `perform` method
-    # - and an isolated container for related private methods
-
-    def perform(curl, mime=nil)
-      # TODO: Is this actually working?
-      # curl.on_missing { |data| puts "4xx: #{data}" }
-      # curl.on_failure { |data| puts "5xx: #{data}" }
-      curl.verbose = @ci.verbose
-      curl.headers['Authorization'] = "Bearer #{@ci.access_token}"
-      curl.headers['Content-Type'] = mime if mime
-      curl.perform
-    end
-  end
-
-  class Downloader < CiClient #:nodoc:
+  class Downloader < SonyCiClient #:nodoc:
     @@cache = {}
 
     def initialize(ci)
