+++ sony-ci-api.gemspec
@@ -1,6 +1,6 @@
 Gem::Specification.new do |s|
   s.name        = 'sony-ci-api'
-  s.version     = '0.0.2'
+  s.version     = '0.1.0'
   s.executables << 'sony-ci-api'
   s.date        = '2015-05-28'
   s.summary     = 'Sony Ci API'
